Cleaned up orientation and implemented RotationSystem

+++ src/main/scala/apsu/demo/rocks/components/AngularVelocity.scala
@@ -5,4 +5,4 @@ package apsu.demo.rocks.components
  *
  * @author david
  */
-case class AngularVelocity(theta: Float)
+case class AngularVelocity(theta: Double)
